meme/requirements.txt
astroid==3.3.8
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
Flask==3.1.0
idna==3.10
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
mccabe==0.7.0
numpy==2.2.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pylint==3.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
requests==2.32.3
six==1.17.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
